requirements.txt
openai-whisper
moviepy
transformers
nltk
pysrt
scenedetect